fix: 오류 수정 및 style 수정
1. list placeholder 출력
2. code 컴포넌트 style 변경
3. 기타 오류 수정
