+++ src/app/playground/page.tsx
@@ -10,7 +10,7 @@ const playground = () => {
   return (
     <>
       <Header page={breadcrumbs} />
-      <div className="flex flex-1 flex-col gap-4 p-2 pt-1">
+      <div className="flex flex-1 flex-col p-2 pt-1">
         <Playground />
       </div>
     </>
